// src/app/models/project/project_responsible.ts
export class ProresponsibleModel {
    constructor() {
             
    }
       
    proresponsible_id: string = "1";
    proresponsible_ref: string = "";
    proresponsible_emp: string = "";   
    proresponsible_position: string = "";   
    proresponsible_area: string = "";   
    proresponsible_fromdate: Date = new Date();
    proresponsible_todate: Date = new Date();
    project_code: string = "";   
    modified_by: string = "";
    modified_date: Date = new Date();
    index: number = 0;
    select: boolean = false;    
   

  }
  